Refactor inference and configuration settings
- Updated the `infer` function to remove the prompt parameter and allow continuous user input until exit commands are given.
- Increased `max_new_tokens` to 2048 for generating longer outputs.
- Modified tokenizer path to use a JSON vocabulary file.
- Adjusted configuration parameters: increased `vocab_size`, `ffn_dim`, and `rope_theta`, while modifying `batch_size`, `grad_accum`, and `total_steps` for improved training efficiency.
- Changed optimizer to "8bit" for better memory management.
- Removed unused model file `evolix.model`.
- Disabled flash SDP in the engine base for compatibility.
- Simplified loss calculation in the trainer by removing unnecessary autocasting.
- Cleaned up the architecture model by removing redundant weight assignments and whitespace.
- Fixed output tensor reshaping in the multi-head attention layer.

+++ src/evolix/models/layers.py
@@ -104,7 +104,7 @@ class MultiHeadLatentAttention(nn.Module):
         with sdpa_kernel(backends=BACKENDS):
             out = F.scaled_dot_product_attention(q_attn, k_attn, v_attn, is_causal=True, dropout_p=0.0, scale=self.scale)
 
-        out = out.transpose(1, 2).reshape(B, T, C).to(x.dtype)
+        out = out.transpose(1, 2).reshape(B, T, C)
         return self.out_proj(out), None
 
     def _forward_decode(self, x, offset: int, kv_cache: Tuple[torch.Tensor, torch.Tensor]):
